Extract wildcard matching into separate class
